Fix issue #14 : [RipsComplex] max_edge_length documented with type 'int'
git-svn-id: svn+ssh://scm.gforge.inria.fr/svnroot/gudhi/trunk@3743 636b058d-ea47-450e-bf9e-a15bfbe3eedb

Former-commit-id: 9ef2e20e818fc1c573ce784cd34acb3c6d9970b7

+++ src/cython/cython/rips_complex.pyx
@@ -51,7 +51,7 @@ cdef class RipsComplex:
         """RipsComplex constructor.
 
         :param max_edge_length: Rips value.
-        :type max_edge_length: int
+        :type max_edge_length: float
 
         :param points: A list of points in d-Dimension.
         :type points: list of list of double
